TestDebugReporter: log the same information that is send to analytics
