+++ maestro-cli/src/main/java/maestro/cli/report/TestDebugReporter.kt
@@ -7,6 +7,7 @@ import maestro.Driver
 import maestro.MaestroException
 import maestro.TreeNode
 import maestro.cli.runner.CommandStatus
+import maestro.cli.util.EnvUtils
 import maestro.debuglog.DebugLogStore
 import maestro.debuglog.LogConfig
 import maestro.orchestra.MaestroCommand
@@ -101,17 +102,17 @@ object TestDebugReporter {
             }
             props["version"].toString()
         }
-        val osName = System.getProperty("os.name")
-        val osVersion = System.getProperty("os.version")
-        val architecture = System.getProperty("os.arch")
-
 
         val logger = LoggerFactory.getLogger("MAESTRO")
         logger.info("---- System Info ----")
-        logger.info("Maestro Version: ${appVersion.getOrNull() ?: "Undefined"}")
-        logger.info("OS Name: $osName")
-        logger.info("OS Version: ${osVersion}")
-        logger.info("Architecture: $architecture")
+        logger.info("Maestro Version: ${EnvUtils.CLI_VERSION ?: "Undefined"}")
+        logger.info("OS Name: ${EnvUtils.OS_NAME}")
+        logger.info("OS Version: ${EnvUtils.OS_VERSION}")
+        logger.info("Architecture: ${EnvUtils.OS_ARCH}")
+        logger.info("Java Version: ${EnvUtils.getJavaVersion()}")
+        logger.info("Xcode Version: ${EnvUtils.getXcodeVersion()}")
+        logger.info("Flutter Version: ${EnvUtils.getFlutterVersionAndChannel().first}")
+        logger.info("Flutter Channel: ${EnvUtils.getFlutterVersionAndChannel().second}")
         logger.info("---------------------")
     }
 
@@ -175,4 +176,3 @@ data class FlowDebugMetadata(
     val screenshots: MutableList<ScreenshotDebugMetadata> = mutableListOf(),
     var exception: MaestroException? = null
 )
-
