+++ maestro-cli/src/main/java/maestro/cli/update/Updates.kt
@@ -4,6 +4,7 @@ import maestro.cli.api.ApiClient
 import maestro.cli.api.CliVersion
 import maestro.cli.util.CiUtils
 import maestro.cli.util.EnvUtils
+import maestro.cli.util.EnvUtils.CLI_VERSION
 import maestro.cli.view.red
 import java.nio.file.Paths
 import java.util.Properties
@@ -17,14 +18,7 @@ import kotlin.io.path.readText
 import kotlin.io.path.writeText
 
 object Updates {
-    val OS_NAME: String = System.getProperty("os.name")
-    val OS_ARCH: String = System.getProperty("os.arch")
     val DEVICE_UUID: String
-    val CLI_VERSION: CliVersion? = getVersion().apply {
-        if (this == null) {
-            System.err.println("\nWarning: Failed to parse current version".red())
-        }
-    }
 
     const val BASE_API_URL = "https://api.mobile.dev"
     private val FRESH_INSTALL: Boolean
@@ -92,19 +86,5 @@ object Updates {
             this.future = future
         }
         return future
-    }
-
-    private fun getVersion(): CliVersion? {
-        val props = try {
-            Updates::class.java.classLoader.getResourceAsStream("version.properties").use {
-                Properties().apply { load(it) }
-            }
-        } catch (e: Exception) {
-            return null
-        }
-
-        val versionString = props["version"] as? String ?: return null
-
-        return CliVersion.parse(versionString)
     }
 }
